+++ src/components/post/index.tsx
@@ -1,81 +1,59 @@
 import styled from '@emotion/styled';
 import { useQuery } from '@tanstack/react-query';
-import { useState } from 'react';
 
-import Genres from '@/components/common/Genres';
 import RecommendationPost from '@/components/common/RecommendationPost';
 import { PostInfo } from '@/components/post/types';
 
+import NoContent from '../common/NoContent';
 import MusicListSkeleton from '../common/skeleton/MusicListSkeleton';
 import { getPostFeedData } from './api';
 
-function PostList() {
-  const [genre, setGenre] = useState('');
+interface Props {
+  genre: string;
+}
 
-  const { data: postFeed, isLoading } = useQuery(['postfeed', genre], () => {
-    return getPostFeedData(genre);
-  });
+function PostList({ genre }: Props) {
+  const { data: postFeed, isLoading } = useQuery(['postfeed', genre], () => getPostFeedData(genre));
 
-  const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const selectedGenre = e.target.value;
+  if (isLoading) {
+    return (
+      <>
+        <MusicListSkeleton />
+        <MusicListSkeleton />
+        <MusicListSkeleton />
+      </>
+    );
+  }
 
-    selectedGenre === 'ALL' ? setGenre('') : setGenre(selectedGenre);
-  };
-
-  return (
-    <Container>
-      <Header>
-        <Title>한눈에 보는 추천</Title>
-      </Header>
-      <Genres shouldNeedAll onChange={onChange} />
-      <PostFeedList>
-        {isLoading ? (
-          <>
-            <MusicListSkeleton />
-            <MusicListSkeleton />
-            <MusicListSkeleton />
-          </>
-        ) : (
-          postFeed?.map(({ postId, nickname, music, likeCount, isBattlePossible }: PostInfo) => (
-            <RecommendationPost
-              key={postId}
-              postId={postId}
-              nickname={nickname}
-              music={music}
-              likeCount={likeCount}
-              isBattlePossible={isBattlePossible}
-            />
-          ))
-        )}
-      </PostFeedList>
-    </Container>
+  return postFeed?.length ? (
+    <>
+      {postFeed?.map(({ postId, nickname, music, likeCount, isBattlePossible }: PostInfo) => (
+        <RecommendationPost
+          key={postId}
+          postId={postId}
+          nickname={nickname}
+          music={music}
+          likeCount={likeCount}
+          isBattlePossible={isBattlePossible}
+        />
+      ))}
+    </>
+  ) : (
+    <Wrapper>
+      <NoContent text='추천 글이 없습니다' isImage width={8} />
+    </Wrapper>
   );
 }
 
 export default PostList;
 
-const Container = styled.div`
-  width: 90%;
-  margin: 0 auto;
-  padding: 2rem 0;
-`;
-
-const Header = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin: 0.5rem 0.5rem 2rem;
-`;
-
-const Title = styled.h1`
-  font-weight: 700;
-  font-size: 1.8rem;
-`;
-
-const PostFeedList = styled.div`
+const Wrapper = styled.div`
   display: flex;
   flex-direction: column;
-  gap: 1.5rem;
-  margin-top: 2.5rem;
-  padding-bottom: 8rem;
+  align-items: center;
+  gap: 0.5rem;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
 `;
